moved more functions into own modules
